Refactored GpuShuffleExchangeExec and ParquetCVShims
