+++ sql-plugin/src/main/spark321db/scala/org/apache/spark/rapids/shims/GpuShuffleExchangeExec.scala
@@ -34,35 +34,5 @@ case class GpuShuffleExchangeExec(
     child: SparkPlan,
     shuffleOrigin: ShuffleOrigin)(
     cpuOutputPartitioning: Partitioning)
-  extends GpuShuffleExchangeExecBaseWithMetrics(gpuOutputPartitioning, child)
-      with ShuffleExchangeLike {
-
-  override def otherCopyArgs: Seq[AnyRef] = cpuOutputPartitioning :: Nil
-
-  override val outputPartitioning: Partitioning = cpuOutputPartitioning
-
-  override def numMappers: Int = shuffleDependencyColumnar.rdd.getNumPartitions
-
-  override def numPartitions: Int = shuffleDependencyColumnar.partitioner.numPartitions
-
-  override def getShuffleRDD(partitionSpecs: Array[ShufflePartitionSpec]): RDD[_] = {
-    new ShuffledBatchRDD(shuffleDependencyColumnar, metrics ++ readMetrics, partitionSpecs)
-  }
-
-  // DB SPECIFIC - throw if called since we don't know how its used
-  override def withNewOutputPartitioning(outputPartitioning: Partitioning) = {
-    throw new UnsupportedOperationException
-  }
-
-  override def runtimeStatistics: Statistics = {
-    // note that Spark will only use the sizeInBytes statistic but making the rowCount
-    // available here means that we can more easily reference it in GpuOverrides when
-    // planning future query stages when AQE is on
-    Statistics(
-      sizeInBytes = metrics("dataSize").value,
-      rowCount = Some(metrics("numOutputRows").value)
-    )
-  }
-
-  override def shuffleId: Int = shuffleDependencyColumnar.shuffleId
-}
+  extends GpuShuffleExchangeExecBase(gpuOutputPartitioning,
+    child, shuffleOrigin)(cpuOutputPartitioning)
